requirements.txt
numpy~=1.19.2
scipy~=1.5.2
matplotlib~=3.3.2
omegaconf~=2.0.2
hydra-core~=1.0.3
torch~=1.7.0
